Initial implementation (Very WIP)
There are a few caveats to integrating the CSound library:

1. There's no pkg-config (or any standard way to add the
includes/libs), so I just liberally added the compile/link flags
for linux-y systems to the makefile.

2. Compiling csound scores is pretty slow (~500ms). I would not
recommend compiling scores in real-time. Instead, I made the loading
portion compile each score and cache it in memory. The problem with
is that compiled scores take up a lot of memory, and it wouldn't be
scalable if someone had a really big soundpack for example. The other
solution would be to store pre-compiled scores in binary form, but
at that point why bother with csound?

3. All calls to csoundPerform*() functions block the current thread.
That means the game freezes until the sound effect is done playing,
or breaks in audio if using the buffered/sampled functions.
The good news is that the CSound library has multithreading functions,
but the bad news is that someone has to manage threads :P

It would be very easy to spawn a new thread for each sound effect,
but that is absolutely not scalable and I think there may be
limitations in my audio driver.

The other option is to have a single secondary thread for csound
playback. You'd need to mix the sound effects into a single buffer,
which is fine, but it would be even better if we could just use the
mixing helpers from SDL (at which point, again, why bother with csound?)